+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+**1.0.0.1002** *(March 28, 2018)*
+
+- Added more tests for the generic search.
+- Fixed a bug in the general data class, where the "owner_id" field was not being set.
+
 **1.0.0.1001** *(March 27, 2018)*
 
 - Added some tests for multiple values of generic search terms.